requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
backoff==2.2.1
bidict==0.23.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dateutils==0.6.12
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
duckdb==1.1.0
et-xmlfile==1.1.0
executing==2.1.0
Flask==3.0.3
Flask-Limiter==3.8.0
Flask-SocketIO==5.4.1
fonttools==4.53.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.1.0
gw_dsl_parser==0.1.49.1
h11==0.14.0
httplib2==0.22.0
idna==3.8
importlib_resources==6.4.5
ipykernel==6.29.5
ipylab==1.0.0
ipython==8.27.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kanaries_track==0.0.5
kiwisolver==1.4.7
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.1
panda==0.3.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pygwalker==0.4.9.8
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
pywin32==306
pyzmq==26.2.0
quickjs==1.19.4
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.2
rsa==4.9
segment-analytics-python==2.2.3
setuptools==74.1.2
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.34
sqlglot==25.20.1
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wasmtime==24.0.0
wcwidth==0.2.13
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
